fix: package json
