+++ product_service/package.json
@@ -8,7 +8,9 @@
     "build": "tsc",
     "cdk": "cdk",
     "cdk:deploy": "cdk deploy --profile=IharTsykala23 --require-approval never",
-    "cdk:bootstrap": "cdk bootstrap"
+    "cdk:bootstrap": "cdk bootstrap",
+    "table:products": "ts-node src/services/fillProductsTable.ts",
+    "table:stocks": "ts-node src/services/fillStocksTable.ts"
   },
   "devDependencies": {
     "@types/aws-lambda": "^8.10.115",
